+++ main.cpp
@@ -76,7 +76,7 @@ int main(int argc, char *argv[])
         app.installTranslator(&translator);
     }
 
-    qDebug("Starting main thread: %u", (unsigned int)QThread::currentThreadId());
+    qDebug("Starting main thread: %lu", (unsigned long)QThread::currentThreadId());
 
     // set the organization/application for QSettings to work properly
     app.setOrganizationName("qcma");
